% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{shinyRetention-package}
\alias{shinyRetention}
\alias{shinyRetention-package}
\title{shinyRetention: Retention Exploration}
\description{
Retention data exploration app-as-a-package.
}
\author{
\strong{Maintainer}: Jumping Rivers \email{info@jumpingrivers.com}

}
\keyword{internal}
